Remove SQL support and add VectorStore utility in RAG integration
Refactor RAG (Reading Comprehension for Question Answering) integration by removing SQL support to simplify the framework. This includes deleting SQL retrieval components and revising RAG pipeline components. In addition, a new utility for handling vector stores has been introduced to facilitate rate-limited adding of documents. Some minor alterations were made to the callback handlers as well.
